Add line preservation when changing existing config files
Python regexp doesn't treat the `\r` symbol as part of the line ending.
So when using windows line endings only `\n` will be treated as end of
line. While `\r` would be a part of matched string.

This commit changes regexp to account for that and also update
tests for this change.

Line preservation is done by using `newline=''` in the `open` function.